requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.11
Django==4.1.13
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-mptt==0.15.0
djlint==1.34.1
djongo==1.3.6
dnspython==2.4.2
EditorConfig==0.12.3
factory-boy==3.3.0
Faker==22.2.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.3
jsbeautifier==1.14.11
json5==0.9.14
MarkupSafe==2.1.3
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pycparser==2.21
pymongo==3.12.3
PySocks==1.7.1
pytest==7.4.4
pytest-base-url==2.0.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-html==4.1.1
pytest-metadata==3.0.0
pytest-selenium==4.0.2
pytest-variables==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4
tenacity==8.2.3
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
wsproto==1.2.0
